+++ packages/patternfly-4/react-core/src/components/Button/Button.js
@@ -33,7 +33,7 @@ const propTypes = {
   isDisabled: PropTypes.bool,
   /** Adds focus styling to the button */
   isFocus: PropTypes.bool,
-  /** Adds hove styling to the button */
+  /** Adds hover styling to the button */
   isHover: PropTypes.bool,
   /** Sets button type */
   type: PropTypes.oneOf(Object.values(ButtonType)),