+++ packages/patternfly-4/react-core/src/components/ContextSelector/ContextSelectorItem.d.ts
@@ -1,0 +1,12 @@
+import { FunctionComponent, HTMLProps, ReactType } from 'react';
+
+export interface ContextSelectorItemProps {
+  isDisabled?: boolean;
+  isSelected?: boolean;
+  isHovered?: boolean;
+  onClick?: Function;
+}
+
+declare const ContextSelectorItem: FunctionComponent<ContextSelectorItemProps>;
+
+export default ContextSelectorItem;