+++ packages/patternfly-4/react-core/src/components/TextInput/TextInput.js
@@ -36,7 +36,7 @@ const propTypes = {
   isReadOnly: PropTypes.bool,
   /** Custom flag to show that the input requires an associated id or aria-label. */
   'aria-label': props => {
-    if (!props.id && !props['aria-label']) {
+    if (!props.id && !props['aria-label'] && !props['aria-labelledby']) {
       return new Error('TextInput requires either an id or aria-label to be specified');
     }
     return null;
@@ -52,7 +52,7 @@ const defaultProps = {
   isDisabled: false,
   isReadOnly: false,
   type: 'text',
-  value: null,
+  value: undefined,
   onChange: () => undefined,
   'aria-label': null
 };