+++ packages/patternfly-4/react-inline-edit-extension/src/components/InlineEdit/index.d.ts
@@ -1,0 +1,19 @@
+export {
+  TableEditConfirmation
+} from './constants';
+
+export * from './utils';
+export {
+  default as editableTableBody,
+  EditableTableBody,
+  InlineEditBodyProps,
+  EditConfig,
+  IEditedCellData,
+} from  './editableTableBody'
+
+export {
+  default as editableRowWrapper,
+  EditableRowWrapper,
+  EditableRowWrapperProps,
+  EditableRowWrapperRow
+} from  './editableRowWrapper'