+++ packages/patternfly-4/react-table/src/components/Table/Body.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { Body } from 'reactabular-table';
 import PropTypes from 'prop-types';
 import { TableContext } from './Table';
+import { isRowExpanded } from './utils';
 
 const propTypes = {
   /** Additional classes for table body. */
@@ -18,50 +19,57 @@ const defaultProps = {
   onRowClick: () => undefined
 };
 
+const flagVisibility = rows => {
+  const visibleRows = rows.filter(oneRow => !oneRow.parent || oneRow.isExpanded);
+  if (visibleRows.length > 0) {
+    visibleRows[0].isFirstVisible = true;
+    visibleRows[visibleRows.length - 1].isLastVisible = true;
+  }
+};
+
 class ContextBody extends React.Component {
-  onRow = (row, props) => {
+  onRow = (row, rowProps) => {
     const { onRowClick } = this.props;
     return {
-      isExpanded: row.isExpanded,
-      isOpen: row.isOpen,
-      onClick: event => {
-        if (event.target.tagName !== 'INPUT' && event.target.tagName !== 'BUTTON') {
-          onRowClick(event, row, props);
-        }
+      row,
+      rowProps,
+      onMouseDown: event => {
+        const computedData = {
+          isInput: event.target.tagName !== 'INPUT',
+          isButton: event.target.tagName !== 'BUTTON'
+        };
+        onRowClick(event, row, rowProps, computedData);
       }
     };
   };
 
-  parentsExpanded(parentId) {
-    const { rows } = this.props;
-    return rows[parentId].hasOwnProperty('parent')
-      ? this.parentsExpanded(rows[parentId].parent)
-      : rows[parentId].isOpen;
-  }
+  mapCells = (headerData, row, rowKey) => {
+    // column indexes start after generated optional columns
+    let additionalColsIndexShift = headerData[0].extraParams.firstUserColumnIndex;
 
-  mapCells = (row, rowKey) => {
-    const { headerData } = this.props;
-    let shiftKey = Boolean(headerData[0] && headerData[0].extraParams.onSelect);
-    shiftKey += Boolean(headerData[0] && headerData[0].extraParams.onCollapse);
     return {
       ...(row &&
         (row.cells || row).reduce(
-          (acc, curr, key) => {
-            const currShift = shiftKey;
-            if (curr.props && curr.props.colSpan) {
-              shiftKey += shiftKey + curr.props && curr.props.colSpan - 1;
+          (acc, cell, cellIndex) => {
+            const isCellObject = cell === Object(cell);
+
+            const mappedCell = {
+              [headerData[cellIndex + additionalColsIndexShift].property]: {
+                title: isCellObject ? cell.title : cell,
+                props: {
+                  isVisible: true,
+                  ...(isCellObject ? cell.props : null)
+                }
+              }
+            };
+
+            // increment the shift index when a cell spans multiple columns
+            if (isCellObject && cell.props && cell.props.colSpan) {
+              additionalColsIndexShift += cell.props.colSpan - 1;
             }
             return {
               ...acc,
-              ...{
-                [headerData[currShift + key].property]: {
-                  title: curr.title || curr,
-                  props: {
-                    isVisible: true,
-                    ...curr.props
-                  }
-                }
-              }
+              ...mappedCell
             };
           },
           { id: row.id !== undefined ? row.id : rowKey }
@@ -71,17 +79,21 @@ class ContextBody extends React.Component {
 
   render() {
     const { className, headerData, rows, rowKey, children, onRowClick, ...props } = this.props;
-    const mappedRows =
-      headerData.length !== 0 &&
-      rows.map((oneRow, oneRowKey) => ({
+
+    let mappedRows;
+    if (headerData.length > 0) {
+      mappedRows = rows.map((oneRow, oneRowKey) => ({
         ...oneRow,
-        ...this.mapCells(oneRow, oneRowKey),
-        ...(oneRow.parent !== undefined
-          ? {
-              isExpanded: this.parentsExpanded(oneRow.parent) && rows[oneRow.parent].isOpen
-            }
-          : {})
+        ...this.mapCells(headerData, oneRow, oneRowKey),
+        isExpanded: isRowExpanded(oneRow, rows),
+        isFirst: oneRowKey === 0,
+        isLast: oneRowKey === rows.length - 1,
+        isFirstVisible: false,
+        isLastVisible: false
       }));
+      flagVisibility(mappedRows);
+    }
+
     return (
       <React.Fragment>
         {mappedRows && (
@@ -101,7 +113,7 @@ class ContextBody extends React.Component {
 
 const TableBody = props => (
   <TableContext.Consumer>
-    {({ headerData, rows }) => <ContextBody {...props} headerData={headerData} rows={rows} />}
+    {({ headerData, rows }) => <ContextBody headerData={headerData} rows={rows} {...props} />}
   </TableContext.Consumer>
 );