+++ packages/patternfly-4/react-table/src/components/Table/RowWrapper.d.ts
@@ -1,0 +1,16 @@
+import { UIEventHandler } from 'react';
+
+export interface RowWrapperRow {
+  isOpen: Boolean;
+  isExpanded: Boolean;
+}
+
+export interface RowWrapper {
+  trRef?: Function;
+  onScroll?: UIEventHandler;
+  onResize?: UIEventHandler;
+  row?: RowWrapperRow;
+  rowProps?: Object;
+}
+
+export default RowWrapper;