+++ packages/patternfly-4/react-table/src/components/Table/index.d.ts
@@ -9,10 +9,14 @@ export {
   ICell,
   IRow,
   IRowData,
+  IColumn,
+  IExtra,
   IExtraData,
   IExtraColumnData
 } from './Table';
 export { default as TableHeader, HeaderProps } from './Header';
 export { default as TableBody, TableBodyProps } from './Body';
-export { sortable, headerCol, cellWidth, ISortable, IExtra } from './utils';
-export { SortByDirection } from './SortColumn'; 
+export { RowWrapper, RowWrapperRow } from './RowWrapper';
+export { default as ExpandableRowContent } from './ExpandableRowContent';
+export { sortable, headerCol, cellWidth, ISortable, expandable, isRowExpanded } from './utils';
+export { SortByDirection } from './SortColumn';