+++ packages/patternfly-4/react-table/src/components/Table/index.js
@@ -1,5 +1,8 @@
-export { default as Table, TableGridBreakpoint, TableVariant } from './Table';
+export { default as Table, TableContext, TableGridBreakpoint, TableVariant } from './Table';
 export { default as TableHeader } from './Header';
 export { default as TableBody } from './Body';
-export { sortable, headerCol, cellWidth } from './utils';
+export { default as BodyWrapper } from './BodyWrapper';
+export { default as RowWrapper } from './RowWrapper';
+export { default as ExpandableRowContent } from './ExpandableRowContent';
+export { sortable, headerCol, cellWidth, expandable, isRowExpanded } from './utils';
 export { SortByDirection } from './SortColumn';