+++ packages/patternfly-4/react-table/src/components/Table/utils/decorators/headerCol.js
@@ -1,10 +1,13 @@
 import React from 'react';
 
 export default (id = 'simple-node') => {
-  const headerCol = (value, { rowIndex } = {}) => ({
-    component: 'th',
-    children: <div id={id + rowIndex}>{value.title || value}</div>
-  });
+  const headerCol = (value, { rowIndex } = {}) => {
+    const result = typeof value === 'object' ? value.title : value;
+    return {
+      component: 'th',
+      children: <div id={id + rowIndex}>{result}</div>
+    };
+  };
 
   return headerCol;
 };