+++ packages/patternfly-4/react-table/src/components/Table/utils/transformers.js
@@ -2,7 +2,7 @@ export { default as selectable } from './decorators/selectable';
 export { default as sortable } from './decorators/sortable';
 export { default as cellActions } from './decorators/cellActions';
 export { default as cellWidth } from './decorators/cellWidth';
-export { collapsible, expandedRow } from './decorators/collapsible';
+export { collapsible, expandedRow, expandable } from './decorators/collapsible';
 export { default as headerCol } from './decorators/headerCol';
 
 export const emptyTD = () => ({